revalidating all paths after tenant name or customDomain change
